requirements.txt
streamlit==1.32.0
pymupdf==1.24.0
sentence-transformers==2.6.1
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
